Added bootstrap data and search bar

+++ src/main/java/com/example/netcracker_lab_2/commands/TrackCommand.java
@@ -14,6 +14,6 @@ public class TrackCommand {
     String name;
     String author;
     String album;
-    Double duration;
+    Integer duration;
     GenreCommand genreCommand;
 }
